Move authentication backend class to separate module.
